+++ generators/swift/model/src/helpers/StructGenerator.ts
