+++ generators/swift/model/src/helpers/index.ts
@@ -1,2 +1,2 @@
 export * from "./ir-types";
-export * from "./struct-generator";
+export * from "./StructGenerator";
